Implement custom hook for fetching secret references
